requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-htmx==1.17.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.7.0